pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.black]
line-length = 120

[tool.isort]
profile = "black"

[tool.poetry]
authors = ["jhcho <jo4186@naver.com>"]
description = ""
name = "projecttemplate-fastapi-dependency-injector"
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
alembic = "^1.13.1"
annotated-types = "0.5.0"
anyio = "3.7.1"
async-timeout = "4.0.3"
asyncpg = "0.28.0"
attrs = "23.1.0"
bcrypt = "4.0.1"
black = {extras = ["d"], version = "^23.11.0"}
bs4 = "^0.0.1"
celery = "^5.3.6"
certifi = "2023.7.22"
cffi = "1.15.1"
click = "8.1.7"
colorama = "0.4.6"
cryptography = "41.0.3"
dependency-injector = "^4.41.0"
ecdsa = "0.18.0"
fakeredis = "^2.21.1"
fastapi = "0.101.1"
greenlet = "2.0.2"
gunicorn = "21.2.0"
h11 = "0.14.0"
httpcore = "0.17.3"
httptools = "0.6.0"
httpx = "0.24.1"
idna = "3.4"
iniconfig = "2.0.0"
isort = "^5.12.0"
jinja2 = "3.1.2"
markupsafe = "2.1.3"
monkeytype = "^23.3.0"
mypy = "^1.7.1"
mypy-extensions = "1.0.0"
numpy = "^1.26.0"
outcome = "1.2.0"
packaging = "23.1"
pandas = "2.1.0"
pandas-stubs = "^2.1.1.230928"
passlib = "1.7.4"
pathspec = "0.11.2"
pika = "^1.3.2"
platformdirs = "3.10.0"
pluggy = "1.3.0"
pyasn1 = "0.5.0"
pycparser = "2.21"
pydantic = "2.3.0"
pydantic-core = "2.6.3"
pydantic-settings = "^2.1.0"
pyjwt = "2.8.0"
pytest = "7.4.1"
pytest-asyncio = "0.21.1"
pytest-runner = "^6.0.1"
pytest-trio = "0.8.0"
python = "^3.11"
python-dateutil = "2.8.2"
python-dotenv = "1.0.0"
python-jose = "3.3.0"
python-multipart = "^0.0.9"
pytz = "2023.3.post1"
pyvirtualdisplay = "^3.0"
pyyaml = "6.0.1"
redis = "^5.0.1"
requests = "^2.31.0"
rsa = "4.9"
selenium = "^4.16.0"
six = "1.16.0"
sniffio = "1.3.0"
sortedcontainers = "2.4.0"
sqlalchemy = "2.0.20"
starlette = "0.27.0"
toml-sort = "^0.23.1"
trio = "0.22.2"
types-passlib = "^1.7.7.13"
types-python-jose = "^3.3.4.8"
types-pyyaml = "^6.0.12.12"
typing-extensions = "4.7.1"
tzdata = "2023.3"
uvicorn = "0.23.2"
watchfiles = "0.19.0"
webdriver-manager = "^4.0.1"
websockets = "11.0.3"
xlrd = "^2.0.1"
